Fix: add arrows
