+++ README.md
@@ -1,0 +1,14 @@
+# Сайт для beelain
+***
+### О проекте:
+Проект создан для изучения HTML, CSS и JS.
+***
+### Что реализованно:
+* Резиновая вёрстка сайта (**desktop first**)
+* Dropdown menu
+***
+### Используемые технологии:
+* HTML (**БЭМ Nested**)
+* SCSS (**Flexbox** и **Grid Layout**)  
+* JavaScript
+* Webpack
