+++ src/index.html
@@ -35,7 +35,7 @@
                     <div class="header__subtitle_block-menu">               
                         <button data-path="one" class="header__subtitle_button header__subtitle_block">
                             <p class="header__subtitle_text">Тарифы и услуги</p>
-                            <svg class="header__subtitle_arrow" width="16" height="16" fill="none" xmlns="http://www.w3.org/2000/svg"><path fill-rule="evenodd" clip-rule="evenodd" d="m6.03 14.03-1.06-1.06L9.44 8.5 4.97 4.03l1.06-1.06 5.53 5.53-5.53 5.53Z" fill="#000"/></svg>
+                            <i data-index-number="one" data-path="one" class="header__subtitle_arrow"></i>
                         </button>
                         <ul data-target="one" class="header__subtitle_menu">
                             <li class="header__subtitle_item">Домашний Интернет</li>
@@ -48,7 +48,7 @@
                     <div class="header__subtitle_block-menu">
                         <button  data-path="two" class="header__subtitle_button header__subtitle_block">
                             <p class="header__subtitle_text">Телефоны и гаджеты</p>
-                            <svg class="header__subtitle_arrow" width="16" height="16" fill="none" xmlns="http://www.w3.org/2000/svg"><path fill-rule="evenodd" clip-rule="evenodd" d="m6.03 14.03-1.06-1.06L9.44 8.5 4.97 4.03l1.06-1.06 5.53 5.53-5.53 5.53Z" fill="#000"/></svg>
+                            <i data-index-number="two" class="header__subtitle_arrow"></i>
                         </button>
                         <ul data-target="two" class="header__subtitle_menu">
                             <li class="header__subtitle_item">Телефоны</li>
@@ -60,7 +60,7 @@
                     <div class="header__subtitle_block-menu">
                         <button  data-path="three" class="header__subtitle_button header__subtitle_block">
                             <p class="header__subtitle_text">Акции и бонусы</p>
-                            <svg class="header__subtitle_arrow" width="16" height="16" fill="none" xmlns="http://www.w3.org/2000/svg"><path fill-rule="evenodd" clip-rule="evenodd" d="m6.03 14.03-1.06-1.06L9.44 8.5 4.97 4.03l1.06-1.06 5.53 5.53-5.53 5.53Z" fill="#000"/></svg>
+                            <i data-index-number="three" class="header__subtitle_arrow"></i>
                         </button>
                         <ul data-target="three" class="header__subtitle_menu">
                             <li class="header__subtitle_item">Выгодное подключение</li>
